user crud authorization applied

+++ src/main/java/com/dsi/ecommerce/controller/UserController.java
@@ -5,8 +5,10 @@ import com.dsi.ecommerce.exception.UserNotFoundException;
 import com.dsi.ecommerce.model.User;
 import com.dsi.ecommerce.service.MyUserDetail;
 import com.dsi.ecommerce.service.UserService;
+import com.dsi.ecommerce.utility.constants.UserRoles;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -17,6 +19,7 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.security.Principal;
 import java.util.List;
+import java.util.Objects;
 
 @RequestMapping("/users")
 @Controller
@@ -39,21 +42,29 @@ public class UserController {
     @RequestMapping("/{username}")
     public String getUserProfile(@PathVariable("username") String username, Model model,
                                  @AuthenticationPrincipal MyUserDetail currentUser){
-        System.out.println(currentUser.getAuthorities());
-        try {
-            User user = userService.getUserByUsername(username);
-            model.addAttribute("user", user);
-        }
-        catch (UserNotFoundException exception){
-            exception.printStackTrace();
-            //todo: handle UserNotFound
-        }
 
-        return "user_profile";
+        if(Objects.equals(username, currentUser.getUsername()) || currentUser.getAuthorities().contains(new SimpleGrantedAuthority(UserRoles.ADMIN.toString())) ){
+            try {
+                User user = userService.getUserByUsername(username);
+                model.addAttribute("user", user);
+            }
+            catch (UserNotFoundException exception){
+                exception.printStackTrace();
+                //todo: handle UserNotFound
+            }
+
+            return "user_profile";
+        }else  return "redirect:/users/" + currentUser.getUsername();
+
     }
 
     @RequestMapping(value = "/{username}/edit", method = RequestMethod.GET)
-    public String updateUserForm(@PathVariable("username") String username, Model model){
+    public String updateUserForm(@PathVariable("username") String username, Model model,
+                                 @AuthenticationPrincipal MyUserDetail currentUser){
+        if(!Objects.equals(username, currentUser.getUsername())  ) {
+            return "redirect:/users/" + currentUser.getUsername() + "/edit";
+        }
+
         try {
             User user = userService.getUserByUsername(username);
             model.addAttribute("user", user);
@@ -72,11 +83,15 @@ public class UserController {
             User user,
             BindingResult bindingResult,
             Model model,
-            RedirectAttributes redirectAttributes
+            RedirectAttributes redirectAttributes,
+            @AuthenticationPrincipal MyUserDetail currentUser
     ){
         if(bindingResult.hasErrors()){
             redirectAttributes.addFlashAttribute(model);
             return "user_profile";
+        }
+        if(!Objects.equals(username, currentUser.getUsername())  ) {
+            return "redirect:/users/" + currentUser.getUsername() + "/edit";
         }
         try{
             User updatedUser = userService.updateUser(username, user);
@@ -89,13 +104,16 @@ public class UserController {
     }
 
     @RequestMapping(value = "/{username}/delete",method = RequestMethod.GET)
-    public String deleteUser(@PathVariable("username") String username, Model model){
+    public String deleteUser(@PathVariable("username") String username, Model model,@AuthenticationPrincipal MyUserDetail currentUser){
+        if(!Objects.equals(username, currentUser.getUsername()) && !currentUser.getAuthorities().contains(new SimpleGrantedAuthority(UserRoles.ADMIN.toString())) ) {
+            return "redirect:/";
+        }
         try{
             userService.deleteUserByUsername(username);
         }catch (UserNotFoundException exception){
             exception.printStackTrace();
             //todo: handle UserNotFound
         }
-        return "redirect:/users";
+        return "redirect:/";
     }
 }
